feat: Update lotteryDrawSchema to handle first winner and refactor activeLotteriesController
This commit updates the `lotteryDrawSchema.js` file to handle the first winner of a lottery draw. It adds a new field `firstWinner` to the schema, which includes the `ticketId` and `address` of the first winner. Additionally, some unnecessary lines in the schema are removed.

The `activeLotteriesController` in the `apiRoutes.js` file is also refactored to properly handle the first winner. If a first winner exists, their information is included in the response. Otherwise, the `firstWinner` field is set to `null`.

Closes #456